reverse naming of payouts table
